requirements.txt
pyTelegramBotAPI
google-genai==1.11.0
supabase
gunicorn
google-api-core
telegramify-markdown
telegramify-markdown[mermaid]